feat(pcg): add split normals, mesh v2 & editor improvements
This commit implements major PCG engine and Unity editor updates:
- Added split normals computation with cusp angle/face group hard edge detection, supporting Auto/Smooth/Flat shading modes
- Upgraded mesh binary format to v2 with vertex normal support, with backward compatibility for v1 and improved read path bounds checking
- Revised bevel_geometry to return PcgGeometry instead of PcgMeshData to preserve face group topology
- Added Unity Editor Group Viewer for group visualization and per-node geometry stats in the graph editor
- Fixed editor state pollution by replacing the static LastCookResultJson with instance-bound cook data
- Updated spline scene handle group member decoding for large vertex counts
- Updated native interop for the new mesh format and JSON stats transmission
- Updated prebuilt macOS native libraries and added comprehensive test coverage for new functionality
- Added code review checklist and context documentation for the change set

Breaking changes:
- Mesh binary format v2 is incompatible with older v1-only readers
- bevel_geometry now returns PcgGeometry instead of PcgMeshData; calling code must be updated
- Default exclude_groups for BevelMesh was removed; explicitly set exclusion groups if needed

+++ Unity/Assets/PcgPlugin/Editor/Graph/PcgCreateSplineSceneHandles.cs
@@ -1500,7 +1500,7 @@ namespace DJTechEditor.PCG.Graph
             public string name;
             public string domain;
             public int count;
-            public int[] members;
+            public long[] members;
         }
 
         [System.Serializable]
@@ -1514,7 +1514,7 @@ namespace DJTechEditor.PCG.Graph
             public string name;
             public string domain;
             public int count;
-            public int[] members;
+            public long[] members;
         }
 
         private static bool IsGroupDomain(SceneEditDomain domain) =>
@@ -1700,8 +1700,8 @@ namespace DJTechEditor.PCG.Graph
                 Handles.color = new Color(0f, 1f, 0.8f, 0.9f);
                 foreach (var key in group.members)
                 {
-                    int a = (int)((long)key / 1000000);
-                    int b = (int)((long)key % 1000000);
+                    int a = (int)(key / 1000000);
+                    int b = (int)(key % 1000000);
                     if (a < 0 || b < 0 || a >= vertices.Length || b >= vertices.Length)
                         continue;
                     var p0 = l2w.MultiplyPoint(vertices[a]);
@@ -1715,7 +1715,7 @@ namespace DJTechEditor.PCG.Graph
                 Handles.color = new Color(1f, 0.85f, 0f, 0.4f);
                 foreach (var faceIdx in group.members)
                 {
-                    int baseIdx = faceIdx * 3;
+                    int baseIdx = (int)faceIdx * 3;
                     if (triangles == null || baseIdx + 2 >= triangles.Length)
                         continue;
                     int v0 = triangles[baseIdx];
@@ -1732,7 +1732,7 @@ namespace DJTechEditor.PCG.Graph
                 Handles.color = new Color(1f, 0.85f, 0f, 0.9f);
                 foreach (var faceIdx in group.members)
                 {
-                    int baseIdx = faceIdx * 3;
+                    int baseIdx = (int)faceIdx * 3;
                     if (triangles == null || baseIdx + 2 >= triangles.Length)
                         continue;
                     int v0 = triangles[baseIdx];
@@ -1753,7 +1753,7 @@ namespace DJTechEditor.PCG.Graph
                 {
                     if (ptIdx < 0 || ptIdx >= vertices.Length)
                         continue;
-                    var p = l2w.MultiplyPoint(vertices[ptIdx]);
+                    var p = l2w.MultiplyPoint(vertices[(int)ptIdx]);
                     Handles.SphereHandleCap(0, p, Quaternion.identity, 0.02f, EventType.Repaint);
                 }
             }
